+++ src/collections/buildSchema.ts
@@ -1,4 +1,6 @@
 import paginate from 'mongoose-paginate-v2';
+import aggregatePaginate from 'mongoose-aggregate-paginate-v2';
+
 import { Schema } from 'mongoose';
 import { SanitizedConfig } from '../config/types';
 import buildQueryPlugin from '../mongoose/buildQuery';
@@ -23,6 +25,10 @@ const buildCollectionSchema = (collection: SanitizedCollectionConfig, config: Sa
 
   schema.plugin(paginate, { useEstimatedCount: true })
     .plugin(buildQueryPlugin);
+
+  if (collection.versions?.drafts) {
+    schema.plugin(aggregatePaginate);
+  }
 
   return schema;
 };