+++ src/config/schema.ts
@@ -36,7 +36,7 @@ export default joi.object({
   cookiePrefix: joi.string(),
   routes: joi.object({
     admin: joi.string(),
-    api: joi.string().allow(''),
+    api: joi.string(),
     graphQL: joi.string(),
     graphQLPlayground: joi.string(),
   }),